Move API client middleware into sub folder
This is a small restructure before adding some further middleware
to the API client.
