+++ common/lib/api-client/middleware/auth.test.js
@@ -7,7 +7,7 @@ const mockConfig = {
   SECRET: 'secret',
 }
 const auth = proxyquire('./auth', {
-  '../../../config': {
+  '../../../../config': {
     API: mockConfig,
   },
 })
