+++ common/lib/api-client/middleware/index.js
@@ -1,0 +1,7 @@
+const auth = require('./auth')
+const errors = require('./errors')
+
+module.exports = {
+  auth,
+  errors,
+}
